requirements.txt
astroid==2.3.3
autopep8==1.4.4
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
gcloud==0.17.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.16.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jws==0.1.3
jwt==0.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
numpy==1.18.1
oauth2client==3.0.0
pandas==1.0.1
passlib==1.7.2
protobuf==3.11.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.6.6
pyDH==0.1.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
Pyrebase==3.0.27
python-dateutil==2.8.1
python-dotenv==0.10.3
python-jwt==2.0.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
scipy==1.4.1
seaborn==0.10.0
six==1.13.0
starlette==0.13.0
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
